fix: add TLS secure context to migrate-mongo config for CDP environments
