+++ package.json
@@ -35,9 +35,9 @@
     "prestart": "npm run build",
     "start": "NODE_ENV=${NODE_ENV:-production} node --enable-source-maps ./.server",
     "setup:husky": "node -e \"try { (await import('husky')).default() } catch (e) { if (e.code !== 'ERR_MODULE_NOT_FOUND') throw e }\" --input-type module",
-    "migrate:up": "migrate-mongo up -f migrate-mongo-config.cjs",
-    "migrate:down": "migrate-mongo down -f migrate-mongo-config.cjs",
-    "migrate:status": "migrate-mongo status -f migrate-mongo-config.cjs"
+    "migrate:up": "migrate-mongo up -f migrate-mongo-config.js",
+    "migrate:down": "migrate-mongo down -f migrate-mongo-config.js",
+    "migrate:status": "migrate-mongo status -f migrate-mongo-config.js"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.679.0",
