+++ src/helpers/secure-context/secure-context.test.js
@@ -1,4 +1,4 @@
-import { getTrustStoreCerts } from '~/src/helpers/secure-context/get-trust-store-certs.js'
+import { getTrustStoreCerts } from '~/src/secure-context.js'
 
 describe('#getTrustStoreCerts', () => {
   const mockProcessEnvWithCerts = {
